requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
lxml>=4.9.0
requests>=2.28.0
python-dotenv>=1.0.0 